services/ingest-load/requirements.txt
requests
psycopg2-binary
apscheduler
python-dateutil
